beth_health_journey_app/scripts/requirements.txt
# Requirements for lab results processing scripts

# PDF extraction
pdfplumber>=0.9.0    # Better accuracy for structured text
pypdf>=3.15.1        # Fallback PDF extraction

# Natural language processing
nltk>=3.8.1          # For text analysis
scikit-learn>=1.3.0  # For classification
spacy>=3.6.1         # For NER and text processing

# Notion integration
notion-client>=2.0.0 # Official Notion API client

# Data processing
pandas>=2.0.0        # For data manipulation
openpyxl>=3.1.2      # For Excel file support
python-dateutil>=2.8.2 # For date parsing

# Utility
tqdm>=4.66.1         # Progress bars
python-dotenv>=1.0.0 # Environment variable management 